+++ packages/unified_distributor/lib/unified_distributor.dart
@@ -1,3 +1,0 @@
-library unified_distributor;
-
-export 'package:fastforge/fastforge.dart';