+++ packages/fastforge/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.6.12
 
-* refactor: move the `unified_distributor` implementation into `fastforge`
-* remove the runtime dependency on `unified_distributor`
+* refactor: move the distributor implementation into `fastforge` and remove
+  the `unified_distributor` package
 
 ## 0.6.11